feat: update ManualHandlerController to AnnotationController
