+++ slipp/src/main/java/slipp/controller/LogInAndOutController.java
@@ -1,0 +1,45 @@
+package slipp.controller;
+
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.annotation.RequestMethod;
+import slipp.domain.User;
+import slipp.support.db.DataBase;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+@Controller
+public class LogInAndOutController {
+    @RequestMapping(value = "/users/loginForm", method = RequestMethod.GET)
+    public String userLoginForm(HttpServletRequest req, HttpServletResponse resp) {
+        return "/user/login.jsp";
+    }
+
+    @RequestMapping(value = "/users/login", method = RequestMethod.POST)
+    public String userLogin(HttpServletRequest req, HttpServletResponse resp) {
+        String userId = req.getParameter("userId");
+        String password = req.getParameter("password");
+        User user = DataBase.findUserById(userId);
+        if (user == null) {
+            req.setAttribute("loginFailed", true);
+            return "/user/login.jsp";
+        }
+        if (user.matchPassword(password)) {
+            HttpSession session = req.getSession();
+            session.setAttribute(UserSessionUtils.USER_SESSION_KEY, user);
+            return "redirect:/";
+        } else {
+            req.setAttribute("loginFailed", true);
+            return "/user/login.jsp";
+        }
+    }
+
+    @RequestMapping(value = "/users/logout", method = RequestMethod.GET)
+    public String userLogout(HttpServletRequest req, HttpServletResponse resp) {
+        HttpSession session = req.getSession();
+        session.removeAttribute(UserSessionUtils.USER_SESSION_KEY);
+        return "redirect:/";
+    }
+}
