--- slipp/src/main/java/slipp/controller/LoginController.java
@@ -1,30 +1,0 @@
-package slipp.controller;
-
-import slipp.domain.User;
-import slipp.support.db.DataBase;
-import nextstep.mvc.asis.Controller;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
-public class LoginController implements Controller {
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        String userId = req.getParameter("userId");
-        String password = req.getParameter("password");
-        User user = DataBase.findUserById(userId);
-        if (user == null) {
-            req.setAttribute("loginFailed", true);
-            return "/user/login.jsp";
-        }
-        if (user.matchPassword(password)) {
-            HttpSession session = req.getSession();
-            session.setAttribute(UserSessionUtils.USER_SESSION_KEY, user);
-            return "redirect:/";
-        } else {
-            req.setAttribute("loginFailed", true);
-            return "/user/login.jsp";
-        }
-    }
-}
