+++ slipp/src/main/java/slipp/controller/UserController.java
@@ -15,13 +15,64 @@ import javax.servlet.http.HttpServletResponse;
 public class UserController {
     private static final Logger log = LoggerFactory.getLogger(UserController.class);
 
+    @RequestMapping(value = "/users", method = RequestMethod.GET)
+    public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
+        if (!UserSessionUtils.isLogined(req.getSession())) {
+            return "redirect:/users/loginForm";
+        }
+
+        req.setAttribute("users", DataBase.findAll());
+        return "/user/list.jsp";
+    }
+
     @RequestMapping(value = "/users/create", method = RequestMethod.POST)
-    public String create(HttpServletRequest req, HttpServletResponse resp) {
+    public String createUser(HttpServletRequest req, HttpServletResponse resp) {
         User user = new User(req.getParameter("userId"), req.getParameter("password"), req.getParameter("name"),
                 req.getParameter("email"));
         log.debug("User : {}", user);
 
         DataBase.addUser(user);
+        return "redirect:/";
+    }
+
+    @RequestMapping(value = "/users/form", method = RequestMethod.GET)
+    public String userForm(HttpServletRequest req, HttpServletResponse resp) {
+        return "/user/form.jsp";
+    }
+
+    @RequestMapping(value = "/users/profile", method = RequestMethod.GET)
+    public String userProfile(HttpServletRequest req, HttpServletResponse resp) {
+        String userId = req.getParameter("userId");
+        User user = DataBase.findUserById(userId);
+        if (user == null) {
+            throw new NullPointerException("사용자를 찾을 수 없습니다.");
+        }
+        req.setAttribute("user", user);
+        return "/user/profile.jsp";
+    }
+
+    @RequestMapping(value = "/users/updateForm", method = RequestMethod.GET)
+    public String userUpdateForm(HttpServletRequest req, HttpServletResponse resp) {
+        String userId = req.getParameter("userId");
+        User user = DataBase.findUserById(userId);
+        if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
+            throw new IllegalStateException("다른 사용자의 정보를 수정할 수 없습니다.");
+        }
+        req.setAttribute("user", user);
+        return "/user/updateForm.jsp";
+    }
+
+    @RequestMapping(value = "/users/update", method = RequestMethod.POST)
+    public String userUpdate(HttpServletRequest req, HttpServletResponse resp) throws Exception {
+        User user = DataBase.findUserById(req.getParameter("userId"));
+        if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
+            throw new IllegalStateException("다른 사용자의 정보를 수정할 수 없습니다.");
+        }
+
+        User updateUser = new User(req.getParameter("userId"), req.getParameter("password"), req.getParameter("name"),
+                req.getParameter("email"));
+        log.debug("Update User : {}", updateUser);
+        user.update(updateUser);
         return "redirect:/";
     }
 }
